'mtls-ping' view refactored
