+++ backend/device_registry/api_views.py
@@ -1,7 +1,6 @@
 import json
 import logging
 import uuid
-import re
 
 from django.http import HttpResponse
 from django.utils import timezone
@@ -13,7 +12,6 @@ from django.db.models.query import QuerySet
 
 from google.cloud import datastore
 from rest_framework import status
-from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.generics import ListAPIView, DestroyAPIView, CreateAPIView, UpdateAPIView, RetrieveAPIView
@@ -39,37 +37,72 @@ else:
     datastore_client = None
 
 
-def is_mtls_authenticated(request):
-    """
-    Returns the device id if authenticated properly
-    through mTLS.
+class MtlsPingView(APIView):
+    """Endpoint for sending a heartbeat."""
+    permission_classes = [AllowAny]
+    authentication_classes = [MTLSAuthentication]
 
-    This should probably be moved to a permission class.
-    """
+    def get(self, request, *args, **kwargs):
+        device = Device.objects.get(device_id=request.device_id)
+        device.last_ping = timezone.now()
+        device.save(update_fields=['last_ping'])
+        portscan_object, _ = PortScan.objects.get_or_create(device=device)
+        firewallstate_object, _ = FirewallState.objects.get_or_create(device=device)
+        block_networks = portscan_object.block_networks.copy()
+        block_networks.extend(settings.SPAM_NETWORKS)
+        return Response({'policy': firewallstate_object.policy_string,
+                         firewallstate_object.ports_field_name: portscan_object.block_ports,
+                         'block_networks': block_networks})
 
-    if not request.META.get('HTTP_SSL_CLIENT_VERIFY') == 'SUCCESS':
-        return Response(
-            'You shall not pass!',
-            status=status.HTTP_403_FORBIDDEN,
-        )
-    cn_domain = re.match(r'.{1}\.(?P<domain>.*)', settings.COMMON_NAME_PREFIX).groupdict()['domain']
+    def post(self, request, *args, **kwargs):
+        data = request.data
+        device = Device.objects.get(device_id=request.device_id)
+        device.last_ping = timezone.now()
+        device.agent_version = data.get('agent_version')
+        device.save(update_fields=['last_ping', 'agent_version'])
 
-    # @TODO clean up this as it will likely break
-    matchObj = re.match(
-        r'.*CN=(.*.{cn_domain})'.format(cn_domain=cn_domain),
-        request.META.get('HTTP_SSL_CLIENT_SUBJECT_DN'),
-        re.M | re.I
-    )
-    if not matchObj:
-        logging.error('[MTLS-Auth] No valid CN found in header HTTP_SSL_CLIENT_SUBJECT_DN.')
-        return False
+        device_info_object, _ = DeviceInfo.objects.get_or_create(device=device)
+        device_info_object.device__last_ping = timezone.now()
+        device_info_object.device_operating_system_version = data.get('device_operating_system_version')
+        device_info_object.fqdn = data.get('fqdn')
+        device_info_object.ipv4_address = data.get('ipv4_address')
+        device_info_object.device_manufacturer = data.get('device_manufacturer')
+        device_info_object.device_model = data.get('device_model')
+        device_info_object.distr_id = data.get('distr_id')
+        device_info_object.distr_release = data.get('distr_release')
+        device_info_object.selinux_state = data.get('selinux_status', {})
+        device_info_object.app_armor_enabled = data.get('app_armor_enabled')
+        device_info_object.logins = data.get('logins', {})
+        device_info_object.default_password = data.get('default_password')
+        device_info_object.save()
 
-    cn = matchObj.group(1)
-    if cn.endswith(settings.COMMON_NAME_PREFIX):
-        return cn
-    else:
-        logging.error('[MTLS-Auth] CN does not match {}'.format(settings.COMMON_NAME_PREFIX))
-        return False
+        portscan_object, _ = PortScan.objects.get_or_create(device=device)
+        scan_info = data.get('scan_info', [])
+        if isinstance(scan_info, str):
+            scan_info = json.loads(scan_info)
+        # Add missing IP protocol version info.
+        for record in scan_info:
+            if 'ip_version' not in record:
+                ipaddr = IPAddress(record['host'])
+                record['ip_version'] = ipaddr.version
+        portscan_object.scan_info = scan_info
+        portscan_object.netstat = data.get('netstat', [])
+        portscan_object.save()
+        firewall_state, _ = FirewallState.objects.get_or_create(device=device)
+        firewall_rules = data.get('firewall_rules', {})
+        if isinstance(firewall_rules, str):
+            firewall_rules = json.loads(firewall_rules)
+        firewall_state.rules = firewall_rules
+        firewall_state.save()
+
+        if datastore_client:
+            task_key = datastore_client.key('Ping')
+            entity = google_cloud_helper.dicts_to_ds_entities(data, task_key)
+            entity['device_id'] = device.device_id  # Will be indexed.
+            entity['last_ping'] = timezone.now()  # Will be indexed.
+            datastore_client.put(entity)
+
+        return Response({'message': 'pong'})
 
 
 class MtlsRenewCertView(APIView):
@@ -358,86 +391,6 @@ class DeviceListView(ListAPIView):
 
     def get_queryset(self):
         return DeviceInfo.objects.filter(device__owner=self.request.user)
-
-
-@api_view(['GET', 'POST'])
-@permission_classes([AllowAny])
-def mtls_ping_view(request, format=None):
-    """
-    Endpoint for sending a heartbeat.
-    """
-
-    device_id = is_mtls_authenticated(request)
-    if type(device_id) is Response:
-        return device_id
-
-    if not device_id:
-        return Response(
-            'Invalid request.',
-            status=status.HTTP_400_BAD_REQUEST
-        )
-
-    if request.method == 'GET':
-        device_object = Device.objects.get(device_id=device_id)
-        device_object.last_ping = timezone.now()
-        device_object.save(update_fields=['last_ping'])
-        portscan_object, _ = PortScan.objects.get_or_create(device=device_object)
-        firewallstate_object, _ = FirewallState.objects.get_or_create(device=device_object)
-        block_networks = portscan_object.block_networks.copy()
-        block_networks.extend(settings.SPAM_NETWORKS)
-        return Response({'policy': firewallstate_object.policy_string,
-                         firewallstate_object.ports_field_name: portscan_object.block_ports,
-                         'block_networks': block_networks})
-
-    elif request.method == 'POST':
-        data = request.data
-        device_object = Device.objects.get(device_id=device_id)
-        device_object.last_ping = timezone.now()
-        device_object.agent_version = data.get('agent_version')
-        device_object.save(update_fields=['last_ping', 'agent_version'])
-
-        device_info_object, _ = DeviceInfo.objects.get_or_create(device=device_object)
-        device_info_object.device__last_ping = timezone.now()
-        device_info_object.device_operating_system_version = data.get('device_operating_system_version')
-        device_info_object.fqdn = data.get('fqdn')
-        device_info_object.ipv4_address = data.get('ipv4_address')
-        device_info_object.device_manufacturer = data.get('device_manufacturer')
-        device_info_object.device_model = data.get('device_model')
-        device_info_object.distr_id = data.get('distr_id', None)
-        device_info_object.distr_release = data.get('distr_release', None)
-        device_info_object.selinux_state = data.get('selinux_status', {})
-        device_info_object.app_armor_enabled = data.get('app_armor_enabled', None)
-        device_info_object.logins = data.get('logins', {})
-        device_info_object.default_password = data.get('default_password')
-        device_info_object.save()
-
-        portscan_object, _ = PortScan.objects.get_or_create(device=device_object)
-        scan_info = data.get('scan_info', [])
-        if isinstance(scan_info, str):
-            scan_info = json.loads(scan_info)
-        # Add missing IP protocol version info.
-        for record in scan_info:
-            if 'ip_version' not in record:
-                ipaddr = IPAddress(record['host'])
-                record['ip_version'] = ipaddr.version
-        portscan_object.scan_info = scan_info
-        portscan_object.netstat = data.get('netstat', [])
-        portscan_object.save()
-        firewall_state, _ = FirewallState.objects.get_or_create(device=device_object)
-        firewall_rules = data.get('firewall_rules', {})
-        if isinstance(firewall_rules, str):
-            firewall_rules = json.loads(firewall_rules)
-        firewall_state.rules = firewall_rules
-        firewall_state.save()
-
-        if datastore_client:
-            task_key = datastore_client.key('Ping')
-            entity = google_cloud_helper.dicts_to_ds_entities(data, task_key)
-            entity['device_id'] = device_id  # Will be indexed.
-            entity['last_ping'] = timezone.now()  # Will be indexed.
-            datastore_client.put(entity)
-
-        return Response({'message': 'pong'})
 
 
 class CredentialsQSMixin(object):
