+++ backend/device_registry/tests/test_api.py
@@ -1,10 +1,12 @@
 from collections import OrderedDict
 import uuid
 from unittest.mock import patch, mock_open
+import json
 
 from django.urls import reverse
 from django.utils import timezone
 from django.contrib.auth import get_user_model
+from django.conf import settings
 
 from rest_framework.test import APITestCase
 from rest_framework import status
@@ -12,7 +14,7 @@ from rest_framework import serializers
 from rest_framework.exceptions import ErrorDetail
 from rest_framework.authtoken.models import Token
 
-from device_registry.models import Credential, Device, DeviceInfo, Tag
+from device_registry.models import Credential, Device, DeviceInfo, Tag, FirewallState, PortScan
 
 TEST_CERT = """-----BEGIN CERTIFICATE-----
 MIIC5TCCAc2gAwIBAgIJAPMjGMrzQcI/MA0GCSqGSIb3DQEBCwUAMBQxEjAQBgNV
@@ -33,6 +35,16 @@ czUUClEc0OJDMw8PsHyYvrl+jk0JFXgDqBgAutPzSiC+pWL3H/5DO8t/NcccNNlR
 1lti7kgwF5QeRU2eEn3VC2F5JreBMpTkeA==
 -----END CERTIFICATE-----
 """
+
+OPEN_PORTS_INFO = [{"host": "192.168.1.178", "port": 22, "proto": "tcp", "state": "open", "ip_version": 4}]
+
+OPEN_CONNECTIONS_INFO = [
+    {'ip_version': 4, 'type': 'tcp', 'local_address': ['192.168.1.178', 4567],
+     'remote_address': ['192.168.1.177', 5678], 'status': 'open', 'pid': 3425}
+]
+
+TEST_RULES = {'INPUT': [{'src': '15.15.15.50/32', 'target': 'DROP'}, {'src': '15.15.15.51/32', 'target': 'DROP'}],
+              'OUTPUT': [], 'FORWARD': []}
 
 
 def datetime_to_str(value):
@@ -672,3 +684,120 @@ class MtlsRenewCertViewTest(APITestCase):
             self.assertEqual(response.data, 'Unknown error')
             self.assertEqual(Device.objects.count(), 1)
             self.assertEqual(DeviceInfo.objects.count(), 1)
+
+
+class MtlsPingViewTest(APITestCase):
+    def setUp(self):
+        self.device = Device.objects.create(device_id='device0.d.wott-dev.local')
+        self.ping_payload = {
+            'device_operating_system_version': 'linux',
+            'fqdn': 'test-device0',
+            'ipv4_address': '127.0.0.1',
+            'uptime': '0',
+            'distr_id': 'Raspbian',
+            'distr_release': '9.4',
+            'scan_info': OPEN_PORTS_INFO,
+            'netstat': OPEN_CONNECTIONS_INFO,
+            'firewall_rules': TEST_RULES
+        }
+        self.url = reverse('mtls-ping')
+        self.headers = {
+            'HTTP_SSL_CLIENT_SUBJECT_DN': 'CN=device0.d.wott-dev.local',
+            'HTTP_SSL_CLIENT_VERIFY': 'SUCCESS'
+        }
+
+    def test_ping_get_success(self):
+        response = self.client.get(self.url, **self.headers)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.data, {'policy': self.device.firewallstate.policy_string,
+                                             'block_ports': [], 'block_networks': settings.SPAM_NETWORKS})
+
+    def test_pong_data(self):
+        # 1st request
+        response = self.client.get(self.url, **self.headers)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.data, {'block_ports': [], 'block_networks': settings.SPAM_NETWORKS,
+                                             'policy': self.device.firewallstate.policy_string})
+        # 2nd request
+        self.device.portscan.block_ports = [['192.168.1.178', 'tcp', 22, False]]
+        self.device.portscan.block_networks = [['192.168.1.177', False]]
+        self.device.portscan.save(update_fields=['block_ports', 'block_networks'])
+        response = self.client.post(self.url, self.ping_payload, **self.headers)
+        self.assertEqual(response.status_code, 200)
+
+        response = self.client.get(self.url, **self.headers)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.data, {'policy': self.device.firewallstate.policy_string,
+                                             'block_ports': [['192.168.1.178', 'tcp', 22, False]],
+                                             'block_networks': [['192.168.1.177', False]] + settings.SPAM_NETWORKS})
+
+    def test_ping_creates_models(self):
+        devinfo_obj_count_before = DeviceInfo.objects.count()
+        portscan_obj_count_before = PortScan.objects.count()
+        self.client.post(self.url, self.ping_payload, **self.headers)
+        devinfo_obj_count_after = DeviceInfo.objects.count()
+        portscan_obj_count_after = PortScan.objects.count()
+        self.assertEqual(devinfo_obj_count_before, 0)
+        self.assertEqual(portscan_obj_count_before, 0)
+        self.assertEqual(devinfo_obj_count_after, 1)
+        self.assertEqual(portscan_obj_count_after, 1)
+
+    def test_ping_writes_scan_info(self):
+        self.client.post(self.url, self.ping_payload, **self.headers)
+        portscan = PortScan.objects.get(device=self.device)
+        scan_info = portscan.scan_info
+        self.assertListEqual(scan_info, OPEN_PORTS_INFO)
+
+    def test_ping_writes_netstat(self):
+        self.client.post(self.url, self.ping_payload, **self.headers)
+        portscan = PortScan.objects.get(device=self.device)
+        netstat = portscan.netstat
+        self.assertListEqual(netstat, OPEN_CONNECTIONS_INFO)
+
+    def test_ping_distr_info(self):
+        self.client.post(self.url, self.ping_payload, **self.headers)
+        self.assertEqual(self.device.deviceinfo.distr_id, 'Raspbian')
+        self.assertEqual(self.device.deviceinfo.distr_release, '9.4')
+
+    def test_ping_writes_firewall_info_pos(self):
+        self.client.post(self.url, self.ping_payload, **self.headers)
+        firewall_state = FirewallState.objects.get(device=self.device)
+        self.assertDictEqual(firewall_state.rules, TEST_RULES)
+
+    def test_ping_writes_firewall_info_neg(self):
+        ping_payload = {
+            'device_operating_system_version': 'linux',
+            'fqdn': 'test-device0',
+            'ipv4_address': '127.0.0.1',
+            'uptime': '0',
+            'scan_info': OPEN_PORTS_INFO,
+            'netstat': OPEN_CONNECTIONS_INFO,
+            'firewall_rules': {'INPUT': [], 'OUTPUT': [], 'FORWARD': []}
+        }
+        self.client.post(self.url, ping_payload, **self.headers)
+        firewall_state = FirewallState.objects.get(device=self.device)
+        self.assertDictEqual(firewall_state.rules, {'INPUT': [], 'OUTPUT': [], 'FORWARD': []})
+
+    def test_ping_converts_json(self):
+        scan_info = [{
+            "host": "localhost",
+            "port": 22,
+            "proto": "tcp",
+            "state": "open",
+            "ip_version": 4
+        }]
+        firewall_rules = {'INPUT': [], 'OUTPUT': [], 'FORWARD': []}
+        ping_payload = {
+            'device_operating_system_version': 'linux',
+            'fqdn': 'test-device0',
+            'ipv4_address': '127.0.0.1',
+            'uptime': '0',
+            'scan_info': json.dumps(scan_info),
+            'firewall_rules': json.dumps(firewall_rules)
+        }
+
+        self.client.post(self.url, ping_payload, **self.headers)
+        firewall_state = FirewallState.objects.get(device=self.device)
+        portscan = PortScan.objects.get(device=self.device)
+        self.assertListEqual(scan_info, portscan.scan_info)
+        self.assertDictEqual(firewall_rules, firewall_state.rules)
