+++ backend/device_registry/urls.py
@@ -36,9 +36,7 @@ if settings.IS_MTLS_API:
         path('api/{}/sign-csr-test'.format(api_version),
              api_views.MtlsRenewCertView.as_view(),
              name='mtls-sign-device-cert-test'),
-        path('api/{}/ping'.format(api_version),
-             api_views.mtls_ping_view,
-             name='mtls-ping'),
+        path('api/{}/ping'.format(api_version), api_views.MtlsPingView.as_view(), name='mtls-ping'),
         path('api/{}/hello'.format(api_version), api_views.MtlsTesterView.as_view(), name='mtls-tester'),
         path('api/{}/action/<int:action_id>/<str:action_name>'.format(api_version), api_views.ActionView.as_view(),
              name='action'),
